+++ woom/tasks.py
@@ -5,6 +5,7 @@ Task manager
 """
 
 import functools
+import logging
 import os
 import re
 
@@ -12,6 +13,7 @@ import configobj
 
 from . import conf as wconf
 from . import render as wrender
+from . import util as wutil
 from .__init__ import WoomError
 
 thisdir = os.path.dirname(__file__)
@@ -124,8 +126,11 @@ class TaskManager:
             Host instance
         """
         self._configs = []
-        self._config = wconf.load_cfg(CFG_DEFAULT_FILE, CFGSPECS_FILE, interpolation=False)
+        self._config = wconf.load_cfg(CFG_DEFAULT_FILE, CFGSPECS_FILE, interpolation=False, list_values=True)
+        self._config_files = [CFG_DEFAULT_FILE]
         self._host = host
+        self.logger = logging.getLogger(__name__)
+        # self._config_files = []
 
     def load_config(self, cfgfile):
         """Load a user configuration file
@@ -141,7 +146,9 @@ class TaskManager:
         ------
         configobj.ConfigObj
         """
-        cfg = wconf.load_cfg(cfgfile, CFGSPECS_FILE, list_values=False)
+        cfgfile = os.path.abspath(cfgfile)
+        cfg = wconf.load_cfg(cfgfile, CFGSPECS_FILE, interpolation=False, list_values=True)
+        self._config_files.append(cfgfile)
         self._configs.append(cfg)
         self._postproc_()
 
@@ -173,7 +180,20 @@ class TaskManager:
                 for artifact_name in content["artifacts"].scalars:
                     path = content["artifacts"][artifact_name]
                     del content["artifacts"][artifact_name]
-                    content["artifacts"][artifact_name] = {"paths": [path], "check": True, "callable": False}
+                    content["artifacts"][artifact_name] = {"path": path, "check": True, "callable": False}
+
+    def __str__(self):
+        return os.pathsep.join(self._config_files)
+
+    def to_json_entry(self):
+        return self._config_files
+
+    @classmethod
+    def from_config_files(cls, host, *config_files):
+        taskmanager = cls(host)
+        for config_file in config_files:
+            taskmanager.load(config_file)
+        return taskmanager
 
     @property
     def config(self):
@@ -220,6 +240,7 @@ class Task:
         self._config = taskconfig
         self._host = host
         self._context = None
+        self.logger = logging.getLogger(__name__)
 
     @property
     def config(self):
@@ -236,6 +257,9 @@ class Task:
         """The task name (:class:`str`)"""
         return self.config.name
 
+    def __str__(self):
+        return self.name
+
     @property
     def is_blocking(self):
         """It is blocking?"""
@@ -248,7 +272,8 @@ class Task:
         ----------
         context: woom.context.Context
         """
-        context["task"] = self
+        if context is not None:
+            context["task"] = self
         self._context = context
 
     @property
@@ -266,6 +291,10 @@ class Task:
     @context.deleter
     def context(self):
         self._context = None
+
+    def has_context(self):
+        """Is the :attr:`context` already set for this task"""
+        return self._context is not None
 
     @functools.cached_property
     def env(self):
@@ -303,12 +332,12 @@ class Task:
         """
         artifacts = {}
         for name, specs in self.config["artifacts"].items():
-            artifacts[name] = self.get_artifact_paths(name)
+            artifacts[name] = self.get_artifact_path(name)
         return artifacts
 
     @functools.lru_cache
-    def get_artifact_paths(self, name):
-        """Get the paths of a single artifact from its name
+    def get_artifact_path(self, name):
+        """Get the path of a single artifact from its name
 
         Parameters
         ----------
@@ -317,33 +346,26 @@ class Task:
 
         Returns
         -------
-        list(str)
-            Lists of raw artifact paths
+        str or list(str)
+            A single pr a lists of raw artifact paths
         """
-        specs = self.config["artifacts"][name]
-        if not isinstance(specs["paths"], list):
-            specs["paths"] = [specs["paths"]]
+        specs = self.config["artifacts"][name].dict()
+        # if not isinstance(specs["path"], list):
+        #     specs["paths"] = [specs["paths"]]
         if specs["callable"]:
-            func_name = specs["paths"][0]
+            func_name = specs["path"]
+            if isinstance(func_name, list):
+                func_name = func_name[0]
             if func_name not in ARTIFACTS_GENERATORS:
                 raise TaskError(f"Artifact generator function not found: {func_name}")
             kwargs = dict(self.context)
             kwargs.update(specs["kwargs"])
-            paths = ARTIFACTS_GENERATORS[func_name](**kwargs)
-            if isinstance(paths, str):
-                paths = [paths]
+            path = ARTIFACTS_GENERATORS[func_name](**kwargs)
+            # if isinstance(paths, str):
+            #     paths = [paths]
         else:
-            paths = specs["paths"]
-        return paths
-
-    # def get_run_dir(self):
-    #     """Get the run directory without rendering"""
-    #     run_dir = self.config["content"]["run_dir"]
-    #     if run_dir is None:
-    #         return ""
-    #     if run_dir == "current":
-    #         run_dir = os.getcwd()
-    #     return run_dir.strip()
+            path = specs["path"]
+        return path
 
     @functools.cached_property
     def run_dir(self):
@@ -377,11 +399,13 @@ class Task:
         if not self.artifacts:
             return {}
         artifacts = {}
-        for name, paths in self.get_artifacts().items():
+        for name, path in self.get_artifacts().items():
+            single = isinstance(path, str)
             artifacts[name] = []
-            for path in paths:
-                rendered = wrender.render(path.strip(), self.context)
-                if not os.path.isabs(path):
+            paths = [path] if single else path
+            for path_ in paths:
+                rendered = wrender.render(path_.strip(), self.context)
+                if not os.path.isabs(path_):
                     if self.run_dir:
                         rendered = os.path.join(self.run_dir, rendered)
                     else:
@@ -389,7 +413,10 @@ class Task:
                             f"Rendered artifact '{name}' of task '{self.name}' is not absolute "
                             "and task run_dir is not defined. Please fix it!"
                         )
-                artifacts[name].append(rendered)
+                if single:
+                    artifacts[name] = rendered
+                else:
+                    artifacts[name].append(rendered)
         return artifacts
 
     def render_content(self):
@@ -406,6 +433,27 @@ class Task:
         # context["context"] = context
         template = wrender.JINJA_ENV.get_template(self.config["content"]["template"])
         return wrender.render(template, self.context)
+
+    def fill_templates(self, dry=False):
+        """Fill static user template files"""
+        for name in self.config["fill"].sections:
+            config = self.config["fill"][name]
+
+            # Render paths
+            template_file = wrender.render(config["template"], self.context)
+            destination = wrender.render(config["destination"], self.context)
+
+            # Fill template
+            self.logger.debug(f"Fill template '{name}': {template_file} → {destination}")
+            template = wrender.JINJA_ENV.get_template(template_file)
+            content = wrender.render(template, self.context)
+
+            # Write destination
+            destination = wutil.check_dir(destination, dry=dry, logger=self.logger)
+            if not dry:
+                with open(destination, "w") as f:
+                    f.write(content)
+            self.logger.info(f"Filled ttemplate '{name}': {template_file} → {destination}")
 
     def export_scheduler_options(self):
         """Export a dict of scheduler options