+++ CHANGES.rst
@@ -7,6 +7,7 @@ Develop
 
 New features
 ------------
+* Add colors to ``woom show status``.
 * App name is now inferred from ``workflow_dir`` when undefined.
 * Add the ``nnodes``, ``ncpus``, ``nnodes`` and ``pmem`` submission options to tasks.
 * Add the ``Context`` class that holds the input mappings for Jinja substitutions [:pull:`16`].
@@ -16,6 +17,8 @@ New features
 * Add a `template` and `blocking` option to task configuration [:pull:`18`].
 * Add a `graceful` option to ``Workflow.kill`` to terminate jobs without killing them [:pull:`18`].
 * Add a sentinel job that monitor jobs when using a scheduler [:pull:`18`].
+* Add the capability to fill templates with a decated section in tasks and with the ``woom fill`` command [:pull:`22`].
+* User paramaters specified in the workflow configuration can now contain sub-sections [:pull:`22`].
 
 Breaking changes
 ----------------
@@ -23,6 +26,7 @@ Breaking changes
 * ``Workflow.get_submission_dir`` is renamed `Workflow.get_task_submission_dir``
 * ``Workflow.get_run_dir`` is renamed ``Workflow.get_task_run_dir``.
 * ``--update `` run option is renamed ``--force`` [:pull:`18`].
+* User paramaters used in jinja rendering are now only accessible in the ``params`` variable [:pull:`22`].
 
 Deprecations
 ------------
@@ -30,6 +34,7 @@ Deprecations
 Bug fixes
 ---------
 * Fix ``Host.get_env`` that was ignoring ``raw_text`` and ``uv_venv`` configuration options.
+* ``run_dir``, ``submission_dir``, ``script_path``, ``env`` must now be prefixed with ``task_`` in a jinja rendering.
 
 Documentation
 -------------